Use "\x00" as C's null string instead of "".
It's a hard problem. In Go there is no nil string. In C there is. And in GTK API
there is a lot of places where nil string is valid. I decided to use special
invalid C string as nil string. It's the shortest valid Go invalid C string:
"\x00". And there is a constant for it: gobject.NilString. Use value or const it
doesn't matter.